PII Review support APIs
- Moved ticketing/progress to person-event/progress.
- Added pii_{started,finished}_at person_event columns to support PII review.
- New config/dashboard-period api to obtain the dashboard period.
